Made changes from front page such that we navigate to the login page rather than the sign up page + made similar chnage to login page as the register page such that if we're already logged in we get redirected to the exercise page
